Remove explicit 3D Secure Force
Stripe’s new Checkout and the Payment Intents API natively support 3D Secure
and trigger it automatically when required (Strong Customer Authentication, etc).

It is also possible to use Radar rules to control when customers are prompted
to complete 3D Secure authentication.

This is preferred way to force 3D secure.