Card form: disable form when pi can't be created
During CC form initialization stripe api is called to create payment
intent. This call can fail with exception, for example when total amount
is less then 0.5 EUR. This exception was not handled which resulted in
payment page displaying 500 server error page.

This commit fixes the situation by catching the exception, disabling CC
form, and displaying reason why it was disabled.

Closes #38